app: add input debouncing for table selection (#1)
Not realizing that the table name was selected, I'd hold down on command launch.
Down wraps to the top table; the only one in my large test database.
Which then queues up a refetch and rerender. The buffer size for this
appears large and makes the app hang for tens of seconds.

Individually, the fetch and render are still fast.

I added a debounce function to determine if the same table is being
fetched again. If one wanted to refresh a single table database, this'll
break that behavior by scrolling through names.